Feature/mnz 76 mnz90

+++ api/src/restaurant/restaurant.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Model } from 'mongoose';
 import { CreateRestaurantDTO, RestaurantResponseDTO, UpdateRestaurantDTO } from './dto';
 import { IRestaurant } from './interfaces';
@@ -15,15 +15,14 @@ export class RestaurantService {
 
   /** API */
   /** Create restaurant */
-  create = async (createRestaurantDTO: CreateRestaurantDTO) => {
+  create = async (createRestaurantDTO: CreateRestaurantDTO, id: string) => {
+    if (createRestaurantDTO.status === true) {
 
-    if(createRestaurantDTO.status === 'true'){
-
-      const findStatus = await this.model.findOne({status: true})
-      if(findStatus){
+      const findStatus = await this.model.findOne({ status: true })
+      if (findStatus) {
 
         findStatus.status = false;
-  
+
         await findStatus.save();
 
       }
@@ -31,7 +30,7 @@ export class RestaurantService {
     }
 
     const restaurant = await new this.model({
-      owner: createRestaurantDTO.ownerId,
+      owner: id,
       name: createRestaurantDTO.name,
       description: createRestaurantDTO.description,
       logoUrl: createRestaurantDTO.logoUrl,
@@ -68,32 +67,33 @@ export class RestaurantService {
 
   /** API */
   /** get restaurant by id */
-  getRestaurantById = async (_id: string) => {
+  getRestaurantById = async (owner: string) => {
 
-    const getRestaurant = await this.model.findById({ _id }).populate("menus")
+    const getRestaurant = await this.model.findOne({ owner }).populate("menus");
+
     return getRestaurant
 
   };
 
   /** API */
   /** update restaurant by id */
-  updateRestaurant = async (_id: string, updateRestaurantDto: UpdateRestaurantDTO) => {
-    
-    if(updateRestaurantDto.status === 'true'){
-      
-      const findStatus = await this.model.findOne({status: true})
-      if(findStatus){
+  updateRestaurant = async (owner: string, _id: string, updateRestaurantDto: UpdateRestaurantDTO) => {
+
+    if (updateRestaurantDto.status === 'true') {
+
+      const findStatus = await this.model.findOne({ status: true })
+      if (findStatus) {
 
         findStatus.status = false;
-  
+
         await findStatus.save();
 
       }
 
     }
-    const updateRestaurant = await this.model.findOneAndUpdate({ _id }, {
+    const updateRestaurant = await this.model.findOneAndUpdate({ owner, _id }, {
       name: updateRestaurantDto.name, description: updateRestaurantDto.description,
-      website: updateRestaurantDto.website, phoneNumber: updateRestaurantDto.phoneNumber, status: updateRestaurantDto.status
+      website: updateRestaurantDto.website, phoneNumber: updateRestaurantDto.phoneNumber, status: updateRestaurantDto.status, logoUrl: updateRestaurantDto.logoUrl
     }, { new: true });
 
     return this.sanitizeRestaurant(updateRestaurant);
@@ -102,10 +102,9 @@ export class RestaurantService {
 
   /** API */
   /** delete restaurant by id */
-  deleteRestaurant = async (_id: string) => {
+  deleteRestaurant = async (owner, _id: string) => {
 
-    const deleteRestaurant = await this.model.findOneAndDelete({ _id });
-
+    const deleteRestaurant = await this.model.findOneAndDelete({ owner, _id });
     return deleteRestaurant;
 
   };
